configuration/pyproject.toml
[tool.poetry]
name = "configuration"
version = "0.1.0"
description = "Configuration for Gilman Lab"
authors = ["Joshua Gilman"]

[tool.poetry.dependencies]
python = "^3.8"
python-consul2 = "^0.1.5"
Jinja2 = "^3.0.1"
PyYAML = "^5.4.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
